requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.14
ml_dtypes==0.5.3
multidict==6.7.0
numpy==2.0.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
propcache==0.4.1
protobuf==4.25.8
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scipy==1.13.1
six==1.17.0
sounddevice==0.5.3
twilio==9.8.5
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
zipp==3.23.0
